Добавить реализацию ТЗ практики 9
