+++ src/server/BaseHttpHandler.java
@@ -1,7 +1,8 @@
 package server;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
+import com.google.gson.*;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
 import com.sun.net.httpserver.HttpExchange;
 import managing.InMemoryTaskManager;
 import tasks.Epic;
@@ -10,12 +11,14 @@ import tasks.Task;
 
 import java.io.IOException;
 import java.io.OutputStream;
+import java.lang.reflect.Type;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
 import static server.HttpTaskServer.DEFAULT_CHARSET;
+import static tasks.Task.FORMATTER;
 
 public class BaseHttpHandler {
     protected InMemoryTaskManager tManager;
@@ -121,6 +124,69 @@ public class BaseHttpHandler {
     }
 
     public static void setGson(Gson gson) {
-        HttpTaskServer.gson = gson;
+        BaseHttpHandler.gson = gson;
+    }
+}
+
+class StatusAdapter extends TypeAdapter<StatusPriority> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final StatusPriority status) throws IOException {
+        jsonWriter.value(status.toString());
+    }
+
+    @Override
+    public StatusPriority read(final JsonReader jsonReader) throws IOException {
+        return StatusPriority.valueOf(jsonReader.nextString());
+    }
+}
+
+class DurationAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final Duration duration) throws IOException {
+        jsonWriter.value(duration.toString());
+    }
+
+    @Override
+    public Duration read(final JsonReader jsonReader) throws IOException {
+        return Duration.parse(jsonReader.nextString());
+    }
+}
+
+class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final LocalDateTime localDate) throws IOException {
+        jsonWriter.value(localDate.format(FORMATTER));
+    }
+
+    @Override
+    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
+        return LocalDateTime.parse(jsonReader.nextString(), FORMATTER);
+    }
+}
+
+class EpicConverter implements JsonSerializer<Epic>, JsonDeserializer<Epic> {
+
+    @Override
+    public JsonElement serialize(Epic epic, Type type, JsonSerializationContext context) {
+        JsonObject object = new JsonObject();
+        object.addProperty("title", epic.getTitle());
+        object.addProperty("description", epic.getDescription());
+        object.addProperty("id", epic.getId());
+        object.addProperty("status", "NEW");
+        object.addProperty("haveSubtasks", epic.isHaveSubtasks());
+        object.addProperty("startTime", epic.getStartTime().format(FORMATTER));
+        object.addProperty("duration", epic.getDuration().toString());
+        return object;
+    }
+
+    @Override
+    public Epic deserialize(JsonElement json, Type type, JsonDeserializationContext context) throws JsonParseException {
+        JsonObject object = json.getAsJsonObject();
+        String title = object.get("title").getAsString();
+        String description = object.get("description").getAsString();
+        return new Epic(title, description);
     }
 }
